+++ R/MazamaRollUtils-package.R
@@ -1,0 +1,5 @@
+#' @keywords internal
+#'
+#' @useDynLib MazamaRollUtils, .registration = TRUE
+#' @importFrom Rcpp evalCpp
+"_PACKAGE"